feat: Add fast-path routing, agent plan persistence, and entity state management
Fast-Path Architecture:
- Situation Assessor: Lightweight LLM classification for simple messages (greetings, thanks, pleasantries) bypassing full reasoning with confidence-based routing (~300ms vs ~600ms)
- Focused Prompts: Fast-path response generation for conversational messages using minimal context, reducing LLM calls for routine interactions
- Conversational Fast-Path: Direct response generation for situation types (GREETING, FAREWELL, THANKS, PLEASANTRY) without agent activation

Agent Plan Management:
- Persistent Execution Plans: Agent-owned plans with task dependencies, blocking states, and full lifecycle management (created → executing → blocked → complete/failed/abandoned)
- Plan State Persistence: Plans stored in StateManager with session-scoped isolation, supporting multi-turn task execution
- Plan Lifecycle Hooks: Automatic plan creation, resumption, abandonment on topic change, and completion tracking

State & Entity Management:
- Entity State Model: Unified patient entities (user-stated) + derived entities (system-resolved) with source tracking and confidence scoring
- Script Detector: Language script detection for translation barrier validation

Infrastructure:
- Feature Flags: Runtime feature toggles for gradual rollout and A/B testing
- Observability Enhancements: Enhanced token usage tracking and reasoning chain logging

Note: Unified reasoning, conversation memory, ReAct architecture, closed-loop validation, execution logging, language context, and booking state tracking were previously committed.

+++ src/patient_ai_service/models/observability.py
@@ -99,7 +99,7 @@ class ReasoningDetails(BaseModel):
 
 class PipelineStep(BaseModel):
     """A single step in the orchestrator pipeline."""
-    step_number: int
+    step_number: float  # Changed to float to support decimal steps like 4.5
     step_name: str
     component: str  # "orchestrator", "translation", "reasoning", "agent", etc.
     inputs: Dict[str, Any] = Field(default_factory=dict)
